Refactor AiTriageAndRemediationAPI into AiTriageAPI and AiRemediationAPI
Split the combined class into two dedicated modules — aiTriageAPI.py
(AiTriageAPI) and aiRemediationAPI.py (AiRemediationAPI) — each owning
its own base URL and methods. Removes aiTriageAndRemediationAPI.py and
updates __init__.py accordingly.
